Fixed warnings when building on windows

+++ FrequencyMagnet/Source/PitchInfoParser.cpp
@@ -13,7 +13,7 @@ static float roundFreq(double var)
     // 3766.66 + .5 =3767.16    for rounding off value
     // then type cast to int so value is 3767
     // then divided by 100 so the value converted into 37.67
-    float value = (int)(var * 100 + .5);
+    float value = static_cast<float>(static_cast<int>(var * 100.0 + .5));
     return (float)value / 100;
 }
